requirements.txt
# Bot Telegram
python-telegram-bot==20.3

# Servidor web (si app.py usa Flask)
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0

# Peticiones HTTP
requests==2.32.5
httpx~=0.24.0
anyio==4.10.0
sniffio==1.3.1

# Librerías de soporte
blinker==1.9.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# Herramientas y utilidades
click==8.3.0
colorama==0.4.6
typing_extensions==4.15.0

# Para manejo de datos y fechas
python-dateutil==2.9.0.post0
six==1.17.0
packaging==25.0
pyparsing==3.2.3
numpy==2.3.2

# Para gráficos / reportes (si los usas)
matplotlib==3.10.5
contourpy==1.3.3
cycler==0.12.1
kiwisolver==1.4.9
fonttools==4.59.1
pillow==11.3.0
reportlab==4.2.2

# Plantillas web
Jinja2==3.1.6
MarkupSafe==3.0.2

# Seguridad / sessions
itsdangerous==2.2.0
Werkzeug==3.1.3
